Combine utils with conftest
