+++ tests/test_loaders.py
@@ -8,8 +8,16 @@ import pandas as pd
 import pyarrow as pa
 import numpy as np
 
-from .utils import assert_columnar_equal
 from mapd.ttypes import TTypeInfo, TColumnType
+
+
+def assert_columnar_equal(result, expected):
+
+    for i, (a, b) in enumerate(zip(result, expected)):
+        np.testing.assert_array_equal(a.nulls, b.nulls)
+        np.testing.assert_array_equal(a.data.int_col, b.data.int_col)
+        np.testing.assert_array_equal(a.data.real_col, b.data.real_col)
+        np.testing.assert_array_equal(a.data.str_col, b.data.str_col)
 
 
 class TestLoaders:
